+++ Backend/test/unit/controllers/chapter.test.js
@@ -1,7 +1,7 @@
 const { getMockReq, getMockRes } = require('@jest-mock/express');
-const chapterController = require('../../app/controllers/chapter.controller');
-const Chapter = require('../../app/models/chapter.model');
-jest.mock('../../app/models/chapter.model');
+const chapterController = require('../../../app/controllers/chapter.controller');
+const Chapter = require('../../../app/models/chapter.model');
+jest.mock('../../../app/models/chapter.model');
 
 Chapter.mockImplementation(() => ({
   listAll: jest.fn(),
@@ -29,9 +29,9 @@ describe('Chapter Controller unit tests', () => {
 
   test('should give an error when listing all chapters', async () => {
     const req = getMockReq();
-    Chapter.listAll.mockRejectedValue(new Error('mock error'));
+    Chapter.listAll.mockRejectedValue(mockError);
     await chapterController.getAll(req, res, next);
-    expect(next).toHaveBeenCalledWith(new Error('mock error'));
+    expect(next).toHaveBeenCalledWith(mockError);
     expect(res.send).toHaveBeenCalledTimes(0);
   });
 });
